problem-service/requirements.txt
alembic==1.4.3
amqp==5.0.1
appdirs==1.4.4
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.1.1
celery==5.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.1
distlib==0.3.1
filelock==3.0.12
firebase-admin==4.4.0
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
google-api-core==1.23.0
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
Mako==1.1.3
MarkupSafe==1.1.1
msgpack==1.0.0
mysql-connector-python==8.0.21
pika==1.1.0
prompt-toolkit==3.0.8
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==0.10.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.19
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
virtualenv==20.0.31
wcwidth==0.2.5
Werkzeug==1.0.1
kazoo==2.8.0
pytest==6.2.4
pact-python==1.3.9

